app/requirements.txt
azure-storage-blob 
pandas
numpy
matplotlib
seaborn
mlflow==2.6.0